moved tileneighbors function to tilemapImporter
